// src/lib/types.ts
export enum SupportedIntegrationTypesToListVulnmapTargets {
  GITHUB = 'github',
  GHE = 'github-enterprise',
  BITBUCKET_CLOUD = 'bitbucket-cloud',
  GCR = 'gcr',
  DOCKER_HUB = 'docker-hub',
  GITLAB = 'gitlab',
  AZURE_REPOS = 'azure-repos',
  BITBUCKET_SERVER = 'bitbucket-server',
}

export interface CommandResult {
  fileName: string | undefined;
  exitCode: number;
  message: string | undefined;
}